supported standard number formatting

+++ pubsub/connection.go
@@ -179,7 +179,7 @@ func (c *connection) handleNewBloom(cmd *NewBloom) error {
 	if !cmd.IsParamsValid() {
 		return ErrInvalidFilterParam
 	}
-	filter, err := bloom.New(cmd.MaxElements, cmd.CollisionProb, MaxBytes)
+	filter, err := bloom.New(uint64(cmd.MaxElements), float64(cmd.CollisionProb), MaxBytes)
 	if err != nil {
 		return fmt.Errorf("bloom filter creation failed %w", err)
 	}
